bot-clean/requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
flask>=3.0.0
gunicorn>=21.2.0
